refactor #270: Service와 UseCase로 분리
